backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-decouple
requests
google-auth
google-auth-oauthlib
backboard-sdk
psycopg2-binary
